[Feature/#3] ListAdapter와 DiffUtil 사용
