+++ app/src/main/java/org/android/go/sopt/data/Repo.kt
@@ -1,0 +1,6 @@
+package org.android.go.sopt.data
+
+data class Repo(
+    val name: String,
+    val author: String
+)
